remove repeated code
