+++ frontend/components/CourseListings/Listings.tsx
@@ -8,39 +8,8 @@ import Typography from '@mui/material/Typography';
 import CoordinatorList from './CoordinatorList';
 import { API_ENDPOINT, CourseEvaluationListEntry } from 'utils/api';
 import CustomTheme from '../utils/CustomTheme';
+import TabPanel, {a11yProps} from "../Custom/TabPanel"
 
-interface TabPanelProps {
-  children?: React.ReactNode;
-  index: number;
-  value: number;
-}
-
-function TabPanel(props: TabPanelProps) {
-  const { children, value, index, ...other } = props;
-
-  return (
-    <div
-      role="tabpanel"
-      hidden={value !== index}
-      id={`simple-tabpanel-${index}`}
-      aria-labelledby={`simple-tab-${index}`}
-      {...other}
-    >
-      {value === index && (
-        <Box sx={{ p: 3 }}>
-          <Typography>{children}</Typography>
-        </Box>
-      )}
-    </div>
-  );
-}
-
-function a11yProps(index: number) {
-  return {
-    id: `simple-tab-${index}`,
-    'aria-controls': `simple-tabpanel-${index}`,
-  };
-}
 
 function Listings() {
